:heavy_plus_sign: IError の種類の追加
