+++ src/MCI.Core/Errors/IO/IFileIOError.cs
@@ -1,0 +1,13 @@
+﻿namespace Miharu.Errors.IO
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Text;
+    using System.Threading.Tasks;
+
+    public interface IFileIOError : IError
+    {
+        string Target { get; }
+    }
+}
