+++ src/MCI.Core/MCI.Core.csproj
@@ -57,6 +57,16 @@
     <Compile Include="DayOfWeekExtensions.cs" />
     <Compile Include="Debugs\ILogger.cs" />
     <Compile Include="Debugs\LoggerHelper.cs" />
+    <Compile Include="Errors\ArgumentOutOfRangeError.cs" />
+    <Compile Include="Errors\BufferOverflowError.cs" />
+    <Compile Include="Errors\IO\DirectoryNotFoundError.cs" />
+    <Compile Include="Errors\IO\FileIsBrokenError.cs" />
+    <Compile Include="Errors\IO\FileIsEmptyError.cs" />
+    <Compile Include="Errors\IO\FileNotFoundError.cs" />
+    <Compile Include="Errors\IO\IFileIOError.cs" />
+    <Compile Include="Errors\IO\NotEnoughDiskSpaceError.cs" />
+    <Compile Include="Errors\NoError.cs" />
+    <Compile Include="Errors\UnkownError.cs" />
     <Compile Include="Logging\LoggerHelper.cs" />
     <Compile Include="Logging\LogLevel.cs" />
     <Compile Include="Logging\LogLevelExtensions.cs" />
@@ -67,7 +77,6 @@
     <Compile Include="Either.cs" />
     <Compile Include="EitherExtensions.cs" />
     <Compile Include="Errors\IError.cs" />
-    <Compile Include="Errors\FileIOError.cs" />
     <Compile Include="Errors\IFormatErrors.cs" />
     <Compile Include="Failure.cs" />
     <Compile Include="Future.cs" />
